+++ aofl-js-packages/resource-enumerate/documentation/js/__config/local-api-config.js
@@ -1,3 +1,0 @@
-export default (apiNs, {hostname}) => {
-  return `//${hostname}/apis/dev/v1.0/resource-enumerate.json`;
-};