+++ aofl-js-packages/resource-enumerate/documentation/js/__config/stage-api-config.js
@@ -1,3 +1,0 @@
-export default (apiNS, {hostname}) => {
-  return `//${hostname}/apis/stage/v1.0/resource-enumerate.json`;
-};