+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # aofl-os changelog
 
 
+## v1.0.0-beta.29 (date)
+Features:
+- aofl-js-packages
+  - resource-enumerate
+    - moved tests to the root project
+    - Config object no  longer supports local/stageRegex. Instead environment is passed to constructor.
+
+---
 ## v1.0.0-beta.28 (09-27-2018)
 Features:
 - Refactord and seperated components